requirements.txt
Flask==3.1.0
SpeechRecognition==3.14.1
pyaudio==0.2.14
opencv-python
mediapipe
numpy

Flask
<<<<<<< HEAD
flask-cors
pymongo
python-dotenv
google-generativeai
flask 

blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.2.2
packaging==24.2
pymongo==4.11
python-dotenv==1.0.1
scikit-learn==1.6.1
scipy==1.15.1
threadpoolctl==3.5.0
Werkzeug==3.1.3
dotenv
pymongo
Flask==3.1.0
SpeechRecognition==3.14.1
pyaudio==0.2.14
opencv-python
mediapipe
numpy
google-generativeai
nltk
googletrans
gtts
deep_translator
flask[async]

Flask
flask_sqlalchemy
python-dotenv
google-generativeai
gtts
deep-translator
=======
Flask-Cors
pymongo
python-dotenv
google-generativeai
>>>>>>> 1038ca7102254afc462e6b46760820e902cbe2bf
